# docs/sdk/openapi/bifrost-api.yaml
openapi: 3.1.0
info:
  title: Bifrost Extensions API
  version: 1.0.0-alpha
  description: |
    Intelligent LLM routing, tool routing, and cost optimization API.

    ## Features
    - Model routing with multiple optimization strategies
    - Cost and latency constraints
    - Tool routing for action-to-tool mapping
    - Prompt classification
    - Usage tracking and analytics

  contact:
    name: SmartCP Team
    email: support@smartcp.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.bifrost.ai
    description: Production server

security:
  - bearerAuth: []

paths:
  /v1/route:
    post:
      summary: Route to optimal model
      description: Route a request to the optimal LLM based on strategy and constraints
      operationId: routeToModel
      tags:
        - Routing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoutingRequest'
            examples:
              simple:
                summary: Simple cost-optimized routing
                value:
                  messages:
                    - role: user
                      content: What is Python?
                  strategy: cost_optimized
              constrained:
                summary: With constraints
                value:
                  messages:
                    - role: user
                      content: Analyze this complex system
                  strategy: performance_optimized
                  constraints:
                    max_cost_usd: 0.02
                    max_latency_ms: 1000
      responses:
        '200':
          description: Successful routing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutingResponse'
              examples:
                success:
                  value:
                    model:
                      model_id: gpt-4o-mini
                      provider: openai
                      estimated_cost_usd: 0.0015
                      estimated_latency_ms: 250
                    confidence: 0.92
                    reasoning: "Cost-optimized selection for simple task"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/RoutingError'

  /v1/route/tool:
    post:
      summary: Route to optimal tool
      description: Route an action to the optimal tool from available options
      operationId: routeToTool
      tags:
        - Routing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolRoutingRequest'
      responses:
        '200':
          description: Successful tool routing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolRoutingDecision'
        '400':
          $ref: '#/components/responses/ValidationError'

  /v1/classify:
    post:
      summary: Classify prompt
      description: Classify a prompt for routing optimization
      operationId: classifyPrompt
      tags:
        - Classification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificationRequest'
      responses:
        '200':
          description: Successful classification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationResult'

  /v1/usage:
    get:
      summary: Get usage statistics
      description: Retrieve usage statistics for a date range
      operationId: getUsageStats
      tags:
        - Analytics
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: group_by
          in: query
          schema:
            type: string
            enum: [model, provider, user]
            default: model
      responses:
        '200':
          description: Usage statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageStats'

  /v1/health:
    get:
      summary: Health check
      description: Check API health status
      operationId: healthCheck
      tags:
        - System
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [healthy, degraded, unhealthy]
                  version:
                    type: string
                  router_available:
                    type: boolean

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key for authentication

  schemas:
    Message:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string

    RoutingStrategy:
      type: string
      enum:
        - cost_optimized
        - performance_optimized
        - speed_optimized
        - balanced
        - pareto

    RoutingConstraints:
      type: object
      properties:
        max_cost_usd:
          type: number
          format: float
        max_latency_ms:
          type: number
          format: float
        required_capabilities:
          type: array
          items:
            type: string

    RoutingRequest:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        strategy:
          $ref: '#/components/schemas/RoutingStrategy'
          default: balanced
        constraints:
          $ref: '#/components/schemas/RoutingConstraints'
        context:
          type: object
          additionalProperties: true

    ModelInfo:
      type: object
      required:
        - model_id
        - provider
        - estimated_cost_usd
        - estimated_latency_ms
      properties:
        model_id:
          type: string
        provider:
          type: string
        estimated_cost_usd:
          type: number
          format: float
        estimated_latency_ms:
          type: number
          format: float

    RoutingResponse:
      type: object
      required:
        - model
        - confidence
      properties:
        model:
          $ref: '#/components/schemas/ModelInfo'
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        reasoning:
          type: string
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/ModelInfo'

    ToolRoutingRequest:
      type: object
      required:
        - action
        - available_tools
      properties:
        action:
          type: string
        available_tools:
          type: array
          items:
            type: string
        context:
          type: object
          additionalProperties: true

    ToolRoutingDecision:
      type: object
      required:
        - recommended_tool
        - confidence
      properties:
        recommended_tool:
          type: string
        confidence:
          type: number
          format: float
        reasoning:
          type: string
        alternatives:
          type: array
          items:
            type: string

    ClassificationRequest:
      type: object
      required:
        - prompt
      properties:
        prompt:
          type: string
        categories:
          type: array
          items:
            type: string

    ClassificationResult:
      type: object
      required:
        - category
        - confidence
      properties:
        category:
          type: string
        confidence:
          type: number
          format: float
        subcategories:
          type: array
          items:
            type: string
        complexity:
          type: string
          enum: [simple, moderate, complex]

    UsageStats:
      type: object
      required:
        - total_requests
        - total_cost_usd
        - avg_latency_ms
        - requests_by_model
        - cost_by_model
      properties:
        total_requests:
          type: integer
        total_cost_usd:
          type: number
          format: float
        avg_latency_ms:
          type: number
          format: float
        requests_by_model:
          type: object
          additionalProperties:
            type: integer
        cost_by_model:
          type: object
          additionalProperties:
            type: number

    Error:
      type: object
      required:
        - error
        - error_code
      properties:
        error:
          type: string
        error_code:
          type: string
        details:
          type: object
          additionalProperties: true

  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid routing request"
            error_code: "VALIDATION_ERROR"
            details:
              field: "messages"
              message: "messages cannot be empty"

    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid or missing API key"
            error_code: "AUTHENTICATION_ERROR"

    RoutingError:
      description: Routing operation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "No models meet the specified constraints"
            error_code: "ROUTING_ERROR"
            details:
              attempted_models: 5
              constraint_violations:
                - "max_cost_usd: 0.01"

tags:
  - name: Routing
    description: Model and tool routing operations
  - name: Classification
    description: Prompt classification
  - name: Analytics
    description: Usage tracking and analytics
  - name: System
    description: System health and status